refactor(cli): extract messaging MCP tools into messaging-tools.ts
Split the 20 messaging-related tool registrations (channels, messages,
threads, DMs, group DMs, reactions, search, and inbox) out of the ~720-line
registerAgentRelayTools function into mcp/messaging-tools.ts behind
registerMessagingTools(server, getAgentClient). These handlers only need the
agent client, so the new module takes just that accessor.

Shared zod input shapes (messageResult, identityOverrideInputShape) move to
mcp/tool-shapes.ts, and the messaging-only resolveEmoji helper moves alongside
its tools. agent-relay-mcp.ts drops to 969 lines (from 2215 before this series).

Pure move with no behavior change; registration order and the full 28-tool
surface are preserved.
